Project restructure, Cuda disabled
It was concluded that at this stage of development, Cuda support is not viable.
