+++ cuda/include/test.h
@@ -1,0 +1,2 @@
+
+// int test(int argc, char const *argv[]);
