+++ cuda/src/main.cpp
@@ -1,0 +1,6 @@
+#include "test.h"
+
+int main(int argc, char const *argv[])
+{
+    return test(argc, argv);
+}
