+++ examples/cuda_testing/CmakeLists.txt
@@ -1,0 +1,13 @@
+cmake_minimum_required(VERSION 3.0.0)
+project(cuda_testing VERSION 0.1.0)
+
+include(CTest)
+enable_testing()
+
+add_executable(cuda_testing cuda_testing.cpp)
+
+target_link_libraries(cuda_testing PRIVATE tezma)
+
+set(CPACK_PROJECT_NAME ${PROJECT_NAME})
+set(CPACK_PROJECT_VERSION ${PROJECT_VERSION})
+include(CPack)
