+++ tests/tensortest.cpp
@@ -1,6 +1,7 @@
 #define BOOST_TEST_MODULE test module name
 #include <boost/test/included/unit_test.hpp>
 
+#include "tezma.h"
 #include "tensor.h"
 
 BOOST_AUTO_TEST_CASE(broadcastable1)
