+++ tezma/include/optimizer.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include "tezma.h"
 #include "tensor.h"
 
 namespace tz
